### 2019-12-17 #17 @lwasser
thank you for this submission @morganjwilliams !! this looks like a great fit for pyopensci .  i'll be in touch about next steps. i'm just catching up from AGU!! :)
