### 2019-12-18 f6f31eb
Update setup.py to include python 3.8
