### 2020-01-02 29b47f6
Update stream_log to deal with ipython logger duplication
