### 2020-01-02 bab6e55
Update stream_log
