### 2020-01-06 b894add
Changes towards updating ternary plots for mpltern
