+++ pyrolite/util/plot.py
@@ -26,9 +26,11 @@ import matplotlib.lines
 import matplotlib.patches
 import matplotlib.path
 import matplotlib.collections
+import matplotlib.artist
 from mpl_toolkits.axes_grid1 import make_axes_locatable
-import matplotlib.axes as matax
+import matplotlib.axes
 from matplotlib.transforms import Bbox
+import mpltern.ternary
 from ..util.math import (
     eigsorted,
     nancov,
@@ -65,6 +67,81 @@ __DEFAULT_DISC_COLORMAP__ = plt.cm.tab10
 FONTSIZE = 12
 
 
+def linekwargs(kwargs):
+    """
+    Get a subset of keyword arguments to pass to a matplotlib line-plot call.
+
+    Parameters
+    -----------
+    kwargs : :class:`dict`
+        Dictionary of keyword arguments to subset.
+
+    Returns
+    --------
+    :class:`dict`
+    """
+    kw = subkwargs(
+        kwargs,
+        plt.plot,
+        matplotlib.axes.Axes.plot,
+        matplotlib.lines.Line2D,
+        matplotlib.collections.Collection,
+    )
+    kw.update(**{"alpha": kwargs.get("alpha")})  # issues with introspection for alpha
+    return kw
+
+
+def scatterkwargs(kwargs):
+    """
+    Get a subset of keyword arguments to pass to a matplotlib scatter call.
+
+    Parameters
+    -----------
+    kwargs : :class:`dict`
+        Dictionary of keyword arguments to subset.
+
+    Returns
+    --------
+    :class:`dict`
+    """
+    kw = subkwargs(
+        kwargs,
+        plt.scatter,
+        matplotlib.axes.Axes.scatter,
+        matplotlib.collections.Collection,
+    )
+    kw.update(**{"alpha": kwargs.get("alpha")})  # issues with introspection for alpha
+    return kw
+
+
+def patchkwargs(kwargs):
+    kw = subkwargs(
+        kwargs,
+        matplotlib.axes.Axes.fill_between,
+        matplotlib.collections.PolyCollection,
+        matplotlib.patches.Patch,
+    )
+    kw.update(**{"alpha": kwargs.get("alpha")})  # issues with introspection for alpha
+    return kw
+
+
+def _mpl_sp_kw_split(kwargs):
+    """
+    Process keyword arguments supplied to a matplotlib plot function.
+
+    Returns
+    --------
+    :class:`tuple` ( :class:`dict`, :class:`dict` )
+    """
+    sctr_kwargs = scatterkwargs(kwargs)
+    # c kwarg is first priority, if it isn't present, use the color arg
+    if sctr_kwargs.get("c") is None:
+        sctr_kwargs = {**sctr_kwargs, **{"c": kwargs.get("color")}}
+
+    line_kwargs = linekwargs(kwargs)
+    return sctr_kwargs, line_kwargs
+
+
 def mappable_from_values(values, cmap=__DEFAULT_CONT_COLORMAP__, **kwargs):
     """
     Create a scalar mappable object from an array of values.
@@ -88,6 +165,147 @@ def marker_cycle(markers=["D", "s", "o", "+", "*"]):
         List of markers to provide to matplotlib.
     """
     return itertools.cycle(markers)
+
+
+def replace_with_ternary_axis(ax):
+    """
+    Replace a specified axis with a ternary equivalent.
+
+    Parameters
+    ------------
+    ax : :class:`~matplotlib.axes.Axes`
+
+    Returns
+    ------------
+    tax : :class:`~mpltern.ternary.TernaryAxes`
+    """
+    fig = ax.figure
+    axes = get_ordered_axes(fig)
+    idx = axes.index(ax)
+    tax = fig.add_subplot(*get_axes_index(ax), projection="ternary")
[…]
