### 2020-01-06 8a34a5a
Update plotting methods for init_axes
