### 2020-01-06 c54d878
Update docs examples
