### 2020-01-06 0ca4f86
And notes for conditional prob in util.plot
