### 2020-01-06 c79d30a
Update plot.density
