+++ pyrolite/plot/density/__init__.py
@@ -1,12 +1,21 @@
+"""
+
+Attributes
+----------
+USE_PCOLOR : :class:`bool`
+    Option to use the :func:`matplotlib.pyplot.pcolor` function in place
+    of :func:`matplotlib.pyplot.pcolormesh`.
+"""
 import matplotlib.pyplot as plt
 import numpy as np
 import logging
 
-from ..comp.codata import close
-from ..util.math import on_finite, linspc_, logspc_, linrng_, logrng_, flattengrid
-from ..util.distributions import sample_kde
-from ..util.plot import (
+from ...comp.codata import close
+from ...util.math import on_finite, linspc_, logspc_, linrng_, logrng_, flattengrid
+from ...util.distributions import sample_kde
+from ...util.plot import (
     ternary_heatmap,
+    xy_to_ABC,
     add_colorbar,
     plot_Z_percentiles,
     percentile_contour_values_from_meshz,
@@ -15,19 +24,23 @@ from ..util.plot import (
     __DEFAULT_DISC_COLORMAP__,
     init_axes,
 )
-from ..util.meta import get_additional_params, subkwargs
+from ...util.meta import get_additional_params, subkwargs
+
+from .grid import DensityGrid
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 logger = logging.getLogger(__name__)
 
+USE_PCOLOR = False
 
-def _get_density_methods(ax, use_pcolor=False):
+
+def _get_density_methods(ax):
     if ax.name == "ternary":
         pcolor = ax.tripcolor
         contour = ax.tricontour
         contourf = ax.tricontourf
     else:
-        if use_pcolor:
+        if USE_PCOLOR:
             pcolor = ax.pcolor
         else:
             pcolor = ax.pcolormesh
@@ -52,7 +65,6 @@ def density(
     vmin=0.0,
     shading="flat",
     colorbar=False,
-    use_pcolor=False,
     no_ticks=False,
     **kwargs
 ):
@@ -84,8 +96,6 @@ def density(
     extent : :class:`list`
         Predetermined extent of the grid for which to from the histogram/KDE. In the
         general form (xmin, xmax, ymin, ymax).
-    coverage_scale : :class:`float`, 1.1
-        Scale the area over which the density plot is drawn.
     contours : :class:`list`
         Contours to add to the plot.
     percentiles :  :class:`bool`, `True`
@@ -100,9 +110,6 @@ def density(
         Shading to apply to pcolormesh.
     colorbar : :class:`bool`, False
         Whether to append a linked colorbar to the generated mappable image.
-    use_pcolor : :class:`bool`
-        Option to use the :func:`matplotlib.pyplot.pcolor` function in place
-        of :func:`matplotlib.pyplot.pcolormesh`.
     no_ticks : :class:`bool`
         Option to *suppress* tickmarks and labels.
 
@@ -137,7 +144,7 @@ def density(
 
     ax = init_axes(ax=ax, projection=projection, **kwargs)
 
-    pcolor, contour, contourf = _get_density_methods(ax, use_pcolor=use_pcolor)
+    pcolor, contour, contourf = _get_density_methods(ax)
     background_color = (*ax.patch.get_facecolor()[:-1], 0.0)
 
     if isinstance(cmap, str):
@@ -150,7 +157,6 @@ def density(
     else:
         cbarlabel = "Frequency"
 
-    exp = (coverage_scale - 1.0) / 2
     valid_rows = np.isfinite(arr).all(axis=-1)
     if (arr.size > 0) and valid_rows.any():
         # Data can't be plotted if there's any nans, so we can exclude these
@@ -158,105 +164,48 @@ def density(
 
         if projection is None:  # binary
             x, y = arr.T
[…]
