### 2020-01-06 c4a7231
Remove ternary import
