### 2020-01-06 320f2b6
Update ternary density
