### 2020-01-06 a10e612
Update density example
