### 2020-01-06 fcb5d93
Update ternary density
