### 2020-01-06 7810c60
Update pyroplot ternary tests
