### 2020-01-06 d1706f0
Update stem test
