### 2020-01-06 dae023a
Update logspacing for density grid
