### 2020-01-07 88d4feb
Update pyrolite.plot for color processing
