### 2020-01-08 57dcd61
Update DISC Colormap
