+++ pyrolite/plot/parallel.py
@@ -4,7 +4,7 @@ import matplotlib.pyplot as plt
 import matplotlib.colors
 from pandas.plotting import parallel_coordinates
 from ..util.meta import subkwargs
-from ..util.plot import __DEFAULT_CONT_COLORMAP__, init_axes
+from ..util.plot import DEFAULT_CONT_COLORMAP, init_axes
 
 
 def parallel(
