### 2020-01-08 7158499
Update plot._c, util.plot
