### 2020-01-08 #17 @morganjwilliams
Happy New Year @lwasser! I hope you've had a bit of time to relax and recover from AGU. Just following up on this one, let me know if there's anything to address before proceeding to a submission.
