### 2020-01-08 86f9800
Update plot tests, color handling
