### 2020-01-08 3e60787
Remove plot.tern from docs
