+++ docs/source/api/plot/tern.rst
@@ -1,5 +1,0 @@
-pyrolite\.plot\.tern
--------------------------------
-  .. automodule:: pyrolite.plot.tern
-      :members:
-      :undoc-members:
