### 2020-01-08 766cebb
Update docs examples, tutorial
