### 2020-01-09 927b4ed
Update ternary colorbars
