+++ pyrolite/plot/density/ternary.py
@@ -1,0 +1,203 @@
+import inspect
+import numpy as np
+import matplotlib.tri
+from ...comp.codata import close, inverse_ilr, ilr, alr, inverse_alr
+from ...util.math import flattengrid
+from ...util.distributions import sample_kde
+from ...util.plot import axes_to_ternary, bin_centres_to_edges
+import logging
+
+logging.getLogger(__name__).addHandler(logging.NullHandler())
+logger = logging.getLogger(__name__)
+
+
+def ternary_heatmap(
+    data,
+    bins=20,
+    mode="density",
+    transform=ilr,
+    inverse_transform=inverse_ilr,
+    ternary_min_value=0.0001,  # 0.01%
+    grid_border_frac=0.1,  # 110% range for grid
+    grid=None,
+    **kwargs
+):
+    """
+    Heatmap for ternary diagrams. This invokes a 3D to 2D transform such as a
+    log transform prior to creating a grid.
+
+    Parameters
+    -----------
+    data : :class:`numpy.ndarray`
+        Ternary data to obtain heatmap coords from.
+    bins : :class:`int`
+        Number of bins for the grid.
+    mode : :class:`str`, :code:`{'histogram', 'density'}`
+        Which mode to render the histogram/KDE in.
+    transform : :class:`callable` | :class:`sklearn.base.TransformerMixin`
+        Callable function or Transformer class.
+    inverse_transform : :class:`callable`
+        Inverse function for `transform`, necessary if transformer class not specified.
+    ternary_min_value : :class:`float`
+        Optional specification of minimum values within a ternary diagram to draw the
+        transformed grid.
+    grid_border_frac : :class:`float`
+        Size of border around the grid, expressed as a fraction of the total grid range.
+    grid : :class:`numpy.ndarray`
+        Grid coordinates to sample at, if already calculated. For the density mode,
+        this is a (nsamples, 2) array. For histograms, this is a two-member list of
+        bin edges.
+
+    Returns
+    -------
+    t, l, r : :class:`tuple` of :class:`numpy.ndarray`
+        Ternary coordinates for the heatmap.
+    H : :class:`numpy.ndarray`
+        Histogram/density estimates for the coordinates.
+    data : :class:`dict`
+        Data dictonary with grid arrays and relevant information.
+
+    Notes
+    -----
+
+    Zeros will not render in this heatmap, consider replacing zeros with small values
+    or imputing them if they must be incorporated.
+    """
+    arr = close(data)  # should remove zeros/nans
+    arr = arr[np.isfinite(arr).all(axis=1)]
+
+    if inspect.isclass(transform):
+        # TransformerMixin
+        tcls = transform()
+        tfm = tcls.transform
+        itfm = tcls.inverse_transform
+    else:
+        # callable
+        tfm = transform
+        assert callable(inverse_transform)
+        itfm = inverse_transform
+
+    if grid is None:
+        # minimum bounds on triangle
+        mins = np.max(
+            np.vstack([np.ones(3) * ternary_min_value, np.nanmin(arr, axis=0)]), axis=0
+        )
+        maxs = np.min(
+            np.vstack([1 - np.ones(3) * ternary_min_value, np.nanmax(arr, axis=0)]),
+            axis=0,
+        )
+
+        # let's construct a bounding triangle
[…]
