### 2020-01-09 b4dc867
Update logo tutorial
