+++ docs/source/examples/geochem/shannon.rst
@@ -45,7 +45,7 @@ Shannon Radii
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  1.750 seconds)
+   **Total running time of the script:** ( 0 minutes  1.493 seconds)
 
 
 .. _sphx_glr_download_examples_geochem_shannon.py:
