+++ docs/source/examples/geochem/lambdas_dimreduction.rst
@@ -181,7 +181,7 @@ Element Patterns in Basalts. J Petrology 57, 1463–1508.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  30.397 seconds)
+   **Total running time of the script:** ( 0 minutes  27.240 seconds)
 
 
 .. _sphx_glr_download_examples_geochem_lambdas_dimreduction.py:

[…]
