### 2020-01-09 d9558d4
Update spider color processing
