+++ docs/source/examples/plotting/stem.rst
@@ -89,7 +89,7 @@ discrete observations down a drill hole:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  1.587 seconds)
+   **Total running time of the script:** ( 0 minutes  1.613 seconds)
 
 
 .. _sphx_glr_download_examples_plotting_stem.py:
