+++ docs/source/examples/comp/logratios.rst
@@ -141,7 +141,7 @@ Finally, we can also see where the logratio mean would fall:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  7.611 seconds)
+   **Total running time of the script:** ( 0 minutes  8.219 seconds)
 
 
 .. _sphx_glr_download_examples_comp_logratios.py:
