+++ docs/source/examples/util/sg_execution_times.rst
@@ -5,7 +5,7 @@
 
 Computation times
 =================
-**00:10.922** total execution time for **examples_util** files:
+**00:09.609** total execution time for **examples_util** files:
 
-- **00:07.344**: :ref:`sphx_glr_examples_util_timescale.py` (``timescale.py``)
-- **00:03.578**: :ref:`sphx_glr_examples_util_TAS.py` (``TAS.py``)
+- **00:06.504**: :ref:`sphx_glr_examples_util_timescale.py` (``timescale.py``)
+- **00:03.106**: :ref:`sphx_glr_examples_util_TAS.py` (``TAS.py``)
