+++ docs/source/tutorials/logo.rst
@@ -302,7 +302,7 @@ We can also do this for individual samples, and estimate percentile contours:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  24.537 seconds)
+   **Total running time of the script:** ( 0 minutes  28.217 seconds)
 
 
 .. _sphx_glr_download_tutorials_logo.py:
