+++ docs/source/examples/plotting/density.rst
@@ -300,7 +300,7 @@ are specified:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  21.752 seconds)
+   **Total running time of the script:** ( 0 minutes  20.910 seconds)
 
 
 .. _sphx_glr_download_examples_plotting_density.py:
