+++ docs/source/examples/util/timescale.rst
@@ -242,7 +242,7 @@ be integrated as a :class:`~pyrolite.util.time.Timescale` method, if there's int
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  7.344 seconds)
+   **Total running time of the script:** ( 0 minutes  6.504 seconds)
 
 
 .. _sphx_glr_download_examples_util_timescale.py:
