### 2020-01-09 f99ec3d
Update plot, processing additional color arguments
