### 2020-01-10 65281fd
Add ternary_density tutorial, update other examples
Update examples, tutorials


Update tutorials


Update density tutorial


Update ternary density tutorial
