### 2020-01-10 dc07ad2
Update setup.py dependencies
