### 2020-01-10 0fa427a
Update setup.py extra requirements
