+++ docs/docs_require.txt
@@ -1,1 +1,1 @@
--e .[dev,skl]
+-e .[docs,skl]
