+++ pyrolite/util/general.py
@@ -1,5 +1,4 @@
 import os, sys
-import psutil
 import re
 import time
 import subprocess, shutil
@@ -220,31 +219,3 @@ def extract_zip(zipfile, output_dir):
                 content = zipfile.open(m, "r").read()
                 with open(str(output_dir / name), "wb") as out:
                     out.write(content)
-
-
-def get_process_tree(process, levels_up=1):
-    """
-    Get a process tree from an active process or process ID.
-
-    Parameters
-    -----------
-    process : :class:`int` | :class:`psutil.Process`
-        Process to search for.
-    levels_up : :class:`int`
-        How many levels up the tree to search for parent processes.
-
-    Returns
-    -------
-    :class:`list`
-        List of processes associated with the given process tree.
-    """
-    if isinstance(process, int):
-        top = psutil.Process(process)
-    elif isinstance(process, psutil.Process):
-        top = process
-    for i in range(levels_up):
-        if top.parent() is not None:
-            top = top.parent()
-        else:
-            break
-    return [top, *top.children(recursive=True)]
