### 2020-01-10 832454a
Update ternary density tutorial
Update density docs examples
