+++ docs/source/examples/plotting/sg_execution_times.rst
@@ -5,14 +5,14 @@
 
 Computation times
 =================
-**01:20.622** total execution time for **examples_plotting** files:
+**00:23.580** total execution time for **examples_plotting** files:
 
-- **00:20.910**: :ref:`sphx_glr_examples_plotting_density.py` (``density.py``)
-- **00:20.023**: :ref:`sphx_glr_examples_plotting_spider.py` (``spider.py``)
-- **00:09.438**: :ref:`sphx_glr_examples_plotting_heatscatter.py` (``heatscatter.py``)
-- **00:08.171**: :ref:`sphx_glr_examples_plotting_REE_v_radii.py` (``REE_v_radii.py``)
-- **00:06.837**: :ref:`sphx_glr_examples_plotting_ternary.py` (``ternary.py``)
-- **00:05.694**: :ref:`sphx_glr_examples_plotting_parallel.py` (``parallel.py``)
-- **00:04.145**: :ref:`sphx_glr_examples_plotting_manifold_vis.py` (``manifold_vis.py``)
-- **00:03.789**: :ref:`sphx_glr_examples_plotting_templates.py` (``templates.py``)
-- **00:01.613**: :ref:`sphx_glr_examples_plotting_stem.py` (``stem.py``)
+- **00:23.580**: :ref:`sphx_glr_examples_plotting_density.py` (``density.py``)
+- **00:00.000**: :ref:`sphx_glr_examples_plotting_REE_v_radii.py` (``REE_v_radii.py``)
+- **00:00.000**: :ref:`sphx_glr_examples_plotting_heatscatter.py` (``heatscatter.py``)
+- **00:00.000**: :ref:`sphx_glr_examples_plotting_manifold_vis.py` (``manifold_vis.py``)
+- **00:00.000**: :ref:`sphx_glr_examples_plotting_parallel.py` (``parallel.py``)
+- **00:00.000**: :ref:`sphx_glr_examples_plotting_spider.py` (``spider.py``)
+- **00:00.000**: :ref:`sphx_glr_examples_plotting_stem.py` (``stem.py``)
+- **00:00.000**: :ref:`sphx_glr_examples_plotting_templates.py` (``templates.py``)
+- **00:00.000**: :ref:`sphx_glr_examples_plotting_ternary.py` (``ternary.py``)
