### 2020-01-10 9c76ad8
Reduce ternary density tutorial figsize
Update tutorials
