+++ docs/source/tutorials/ternary_density.rst
@@ -78,7 +78,7 @@ and interpoalted for the ternary heatmap:
     from pyrolite.util.math import flattengrid
     from pyrolite.util.plot import axes_to_ternary, share_axes
 
-    fig, ax = plt.subplots(3, 2, figsize=(8, 12))
+    fig, ax = plt.subplots(3, 2, figsize=(6, 9))
     ax = ax.flat
 
     share_axes([ax[1], ax[2], ax[3]])
@@ -158,7 +158,7 @@ We can see how this works almost exactly the same for the histograms:
 
 .. code-block:: default
 
-    fig, ax = plt.subplots(3, 2, figsize=(8, 12))
+    fig, ax = plt.subplots(3, 2, figsize=(6, 9))
     ax = ax.flat
 
     share_axes([ax[1], ax[2], ax[3]])
@@ -204,7 +204,7 @@ We can see how this works almost exactly the same for the histograms:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  11.744 seconds)
+   **Total running time of the script:** ( 0 minutes  8.762 seconds)
 
 
 .. _sphx_glr_download_tutorials_ternary_density.py:
