### 2020-01-10 81f378a
Update util.plot for ternary stdev ellipses
