### 2020-01-10 feb0a3d
Update logo tutorial
