+++ docs/source/tutorials/logo.rst
@@ -174,6 +174,7 @@ vectors using principal component analysis:
 
 .. code-block:: default
 
+
     kwargs = dict(ax=ax[1], transform=from_log, nstds=3)
     ax[1].set_title("Covariance Ellipses and PCA Vectors")
     for ix, sample in enumerate(df.Sample.unique()):
@@ -302,7 +303,7 @@ We can also do this for individual samples, and estimate percentile contours:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  28.217 seconds)
+   **Total running time of the script:** ( 0 minutes  29.704 seconds)
 
 
 .. _sphx_glr_download_tutorials_logo.py:
