### 2020-01-10 8019301
Remove unused functions from.util.plot
