+++ pyrolite/util/plot.py
@@ -775,161 +775,6 @@ def ternary_grid(
     return bins, binedges, centregrid, edgegrid
 
 
-def set_ternary_labels(tax, labels, fontsize=FONTSIZE, offset=0.2, axlabels=True):
-    """
-    Set the labels and positions in a ternary plot.
-    """
-    if len(tax._labels.keys()) and not axlabels:  # are labels, should be none
-        labels = [None, None, None]
-    tax.right_axis_label(
-        labels[0],
-        fontsize=fontsize,
-        offset=0,
-        position=(1.0, 0, 0),
-        rotation=0.0,
-        verticalalignment="top",
-        # horizontalalignment="left",
-    )
-    tax.left_axis_label(
-        labels[1],
-        fontsize=fontsize,
-        offset=0,
-        position=(-offset / 2, 1.0 + offset, -offset / 2),
-        rotation=0.0,
-        verticalalignment="bottom",
-        # horizontalalignment="center",
-    )
-    tax.bottom_axis_label(
-        labels[2],
-        fontsize=fontsize,
-        offset=0,
-        position=(0.0, 0.0, 1.0),
-        rotation=0.0,
-        verticalalignment="top",
-        # horizontalalignment="right",
-    )
-
-
-def ternarygrid_to_xy(arr, tfm=lambda x: x, yscale=1.0):
-    shape = arr[0].shape
-    flat = np.vstack([a.flatten() for a in arr])
-    xi, yi = ABC_to_xy(tfm(flat.T), yscale=yscale).T
-    xi, yi = xi.reshape(shape), yi.reshape(shape)
-    return xi, yi
-
-
-def ternary_heatmap(
-    data,
-    sample_at=None,
-    bins=10,
-    margin=0.01,
-    force_margin=False,
-    remove_background=True,
-    transform=ilr,
-    inverse_transform=inverse_ilr,
-    mode="histogram",
-    aspect="eq",
-    ret_centres=False,
-    **kwargs
-):
-    """
-    Heatmap for ternary diagrams.
-
-    Parameters
-    -----------
-    data : :class:`numpy.ndarray`
-        Ternary data to obtain heatmap coords from.
-    sample_at : :class:`numpy.ndarray`:
-        Locations to sample the heatmap at rather than a specificed grid.
-    bins : :class:`int`
-        Number of bins for the grid.
-    margin : :class:`float`
-        Optional specification of margin around ternary diagram to draw the grid.
-    force_margin : :class:`bool`
-        Whether to enforce the minimum margin.
-    remove_background : :class:`bool`
-        Whether to display cells with no counts.
-    transform : :class:`callable` | :class:`sklearn.base.TransformerMixin`, :func:`~pyrolite.comp.codata.ilr`
-        Callable function or Transformer class.
-    inverse_transform : :class:`callable`, :func:`~pyrolite.comp.codata.inverse_ilr`
-        Inverse function for `transform`, necessary if transformer class not specified.
-    mode : :class:`str`, {'histogram', 'density'}
-        Which mode to render the histogram/KDE in.
-    aspect : :class:`str`, {'unit', 'equilateral'}
-        Aspect of the ternary plot - whether to plot with an equilateral triangle
-        (yscale = 3**0.5/2) or a triangle within a unit square (yscale = 1.)
-    ret_centres : :class:`bool`
-        Whether to return the centres of the ternary bins.
-
-    Returns
-    -------
-    :class:`tuple` of :class:`numpy.ndarray`
-        :code:`x` bin edges :code:`xe`, :code:`y` bin edges :code:`ye`, histogram/density estimates :code:`Z`.
-        If :code:`ret_centres` is :code:`True`, the last return value will contain the
-        bin :code:`centres`.
-
-    Todo
-    -----
[…]
