### 2020-01-10 618a0ae
Update subplots_kw arguments for axes in util.plot
Update subplot_kw


Test typo
