+++ docs/source/dev/changelog.rst
@@ -14,13 +14,59 @@ All notable changes to this project will be documented here.
 Todo
 ~~~~~~~
 
-* Update ternary plots to use :mod:`mpltern`
-  (`#28 <https://github.com/morganjwilliams/pyrolite/issues/28>`__)
-* Upgrades for :func:`~pyrolite.geochem.transform.convert_chemistry`
 * Index memory for :func:`~pyrolite.plot.spider.spider`
   (`#27 <https://github.com/morganjwilliams/pyrolite/issues/27>`__)
 * Updates to include more lithogeochemical plot templates
   (`#26 <https://github.com/morganjwilliams/pyrolite/issues/26>`__)
+* Upgrades for :func:`~pyrolite.geochem.transform.convert_chemistry`
+
+`0.2.4`_
+--------------
+
+* Removed Python 3.5 support, added Python 3.8 support.
+* Updated ternary plots to use :mod:`mpltern`
+  (`#28 <https://github.com/morganjwilliams/pyrolite/issues/28>`__)
+* Added a
+  `ternary heatmap tutorial <https://pyrolite.readthedocs.io/en/develop/tutorials/ternary_density.html>`__
+
+:mod:`pyrolite.plot`
+~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* Added :meth:`pyrolite.plot.pyroplot.plot` method
+* Removed :meth:`pyrolite.plot.pyroplot.ternary` method (ternary plots now served
+  through the same interface as bivariate plots using
+  :meth:`pyrolite.plot.pyroplot.scatter`, :meth:`pyrolite.plot.pyroplot.plot`,
+  and :meth:`pyrolite.plot.pyroplot.plot`)
+* Added :mod:`pyrolite.plot._c` for processing color arguments.
+* Moved :mod:`pyrolite.plot.density` to its own sub-submodule, including
+  :mod:`pyrolite.plot.density.ternary` and :mod:`pyrolite.plot.density.grid`
+
+:mod:`pyrolite.util`
+~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* Updated :mod:`~pyrolite.util.time` to include official colors.
+* Added :mod:`pyrolite.util.time`
+  `example <https://pyrolite.readthedocs.io/en/develop/examples/util/timescale.html>`__
+* Updated :func:`~pyrolite.util.meta.stream_log` to deal with logger
+  duplication issues.
+* Various updates to :mod:`pyrolite.util.plot`, noted below:
+* Added universal axes initiation for bivariate/ternary diagrams using
+  :func:`~pyrolite.util.plot.init_axes` and axes labelling with
+  :func:`~pyrolite.util.plot.label_axes`,
+* Added keyword argument processing functions :func:`~pyrolite.util.plot.scatterkwargs`,
+  :func:`~pyrolite.util.plot.linekwargs`, and
+  :func:`~pyrolite.util.plot.patchkwargs`
+* Added functions for replacing non-projected axes with ternary axes, including
+  :func:`~pyrolite.util.plot.replace_with_ternary_axis`,
+  :func:`~pyrolite.util.plot.axes_to_ternary` (and
+  :func:`~pyrolite.util.plot.get_axes_index` to maintain ordering of new axes)
+* Added :func:`pyrolite.util.plot.get_axis_density_methods` to access the relevant
+  histogram/density methods for bivariate and ternary axes
+* Renamed private attributes for default colormaps to
[…]
