### 2020-01-10 f8f7fe0
Update plot.color, density docstring
