### 2020-01-10 ddef85b
Update docs structure, logo tutorial
