+++ docs/source/tutorials/logo.rst
@@ -19,7 +19,6 @@ analysis.
     import matplotlib
     import numpy as np
     import pandas as pd
-    import ternary as pyternary
     import matplotlib.pyplot as plt
     import matplotlib.colors
     import matplotlib.cm
@@ -28,16 +27,7 @@ analysis.
     from pyrolite.util.synthetic import random_composition
 
     import pyrolite.plot
-    from pyrolite.util.plot import (
-        plot_pca_vectors,
-        plot_stdev_ellipses,
-        ternary_heatmap,
-        plot_Z_percentiles,
-        percentile_contour_values_from_meshz,
-        bin_centres_to_edges,
-        bin_edges_to_centres,
-        ternary_patch,
-    )
+    from pyrolite.util.plot import plot_pca_vectors, plot_stdev_ellipses
 
     np.random.seed(82)
 
@@ -174,7 +164,6 @@ vectors using principal component analysis:
 
 .. code-block:: default
 
-
     kwargs = dict(ax=ax[1], transform=from_log, nstds=3)
     ax[1].set_title("Covariance Ellipses and PCA Vectors")
     for ix, sample in enumerate(df.Sample.unique()):
@@ -303,7 +292,7 @@ We can also do this for individual samples, and estimate percentile contours:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  29.704 seconds)
+   **Total running time of the script:** ( 0 minutes  29.095 seconds)
 
 
 .. _sphx_glr_download_tutorials_logo.py:
