+++ docs/source/api/plot/density/density_grid.rst
@@ -1,0 +1,5 @@
+pyrolite\.plot\.density\.grid
+-------------------------------
+  .. automodule:: pyrolite.plot.density.grid
+      :members:
+      :undoc-members:
