+++ docs/source/api/plot/density/density_ternary.rst
@@ -1,0 +1,5 @@
+pyrolite\.plot\.density\.ternary
+----------------------------------
+  .. automodule:: pyrolite.plot.density.ternary
+      :members:
+      :undoc-members:
