### 2020-01-10 352c094
Update logo tutorial
