+++ docs/source/tutorials/logo.py
@@ -93,7 +93,7 @@ kwargs = dict(marker="D", alpha=0.2, s=3, no_ticks=True, axlabels=False)
 for ix, sample in enumerate(df.Sample.unique()):
     comp = df.query("Sample == {}".format(sample))
     comp.loc[:, chem].pyroplot.scatter(ax=ax[0], c=t10b3[ix], **kwargs)
-fig
+plt.show()
 #######################################################################################
 # We can take the mean and covariance in log-space to create covariance ellipses and
 # vectors using principal component analysis:
@@ -105,12 +105,12 @@ for ix, sample in enumerate(df.Sample.unique()):
     tcomp = to_log(comp.loc[:, chem])
     plot_stdev_ellipses(tcomp.values, color=t10b3[ix], resolution=1000, **kwargs)
     plot_pca_vectors(tcomp.values, ls="-", lw=0.5, color="k", **kwargs)
-fig
+plt.show()
 #######################################################################################
 # We can also look at data density (here using kernel density estimation)
 # in logratio-space:
 #
-kwargs = dict(ax=ax[-2], bins=100, no_ticks=True, axlabels=False)
+kwargs = dict(ax=ax[-2], bins=100, axlabels=False)
 ax[-2].set_title("Individual Density, with Contours")
 
 for ix, sample in enumerate(df.Sample.unique()):
@@ -122,16 +122,16 @@ for ix, sample in enumerate(df.Sample.unique()):
         contour_labels={0.68: "σ", 0.95: "2σ"},
         **kwargs,
     )
-fig
+plt.show()
 #######################################################################################
 # We can also do this for individual samples, and estimate percentile contours:
 #
-kwargs = dict(ax=ax[-1], no_ticks=True, axlabels=False)
+kwargs = dict(ax=ax[-1], axlabels=False)
 ax[-1].set_title("Overall Density")
 df.loc[:, chem].pyroplot.density(bins=100, cmap="Greys", **kwargs)
-fig
+plt.show()
 #######################################################################################
 for a in ax:
     a.set_aspect("equal")
     a.patch.set_visible(False)
-fig
+plt.show()
