### 2020-01-10 b69de68
Add plot.color api page
