+++ docs/source/api/plot/color.rst
@@ -1,0 +1,5 @@
+pyrolite\.plot\.color
+-------------------------------
+  .. automodule:: pyrolite.plot.color
+      :members:
+      :undoc-members:
