### 2020-01-10 b03a9bb
Minor changelog update
