### 2020-01-14 cdc1862
Add paper.bib
