+++ paper/paper.bib
@@ -1,0 +1,132 @@
+@article{Blundy2003,
+  ids = {Blundy2003a},
+  title = {Partitioning of Trace Elements between Crystals and Melts},
+  author = {Blundy, Jon and Wood, Bernard},
+  year = {2003},
+  month = may,
+  volume = {210},
+  pages = {383--397},
+  issn = {0012-821X},
+  doi = {10.1016/S0012-821X(03)00129-8},
+  abstract = {Advances in analytical geochemistry have made it possible to determine precisely the concentration of many trace elements and their isotopes in rocks. These data provide the cornerstone for geochemical models of the Earth and terrestrial planets. However, our understanding of how trace elements behave has not kept pace with the analytical advances. As a result, geochemists are often hampered in their interpretation of geochemical data by an incomplete knowledge of trace element partitioning under the conditions of interest. Through advances in trace element microbeam analysis it is now possible to determine partition coefficients experimentally under important conditions, such as during melting of the crust and mantle. This large body of experimental data can be used to investigate the fundamental controls on element partitioning. Simple continuum theories of elastic strain and point charges in crystal lattices, that account, respectively, for mismatch in ionic radius and ionic charge between the substituent trace ion and the lattice site on which it is accommodated, provide a very useful theoretical framework. This approach can be used as the basis for quantitative models of trace element partitioning, in terms of pressure, temperature, redox state and composition, and as a means of predicting partition coefficients for elements not routinely analysed. Experimental studies of partitioning are supported by atomistic computer simulations at zero K. Developments in computational techniques that enable direct simulation of high temperature and pressure mineral-melt partitioning will revolutionise the field in the near future. Use of novel, spectroscopic techniques to probe the structural environment of trace elements in crystals and glasses will provide valuable new data for computational and theoretical models. Extension of high temperature partitioning theory to ambient conditions is an essential step in understanding current climate change proxies, and tackling a host of environmental problems.},
+  journal = {Earth and Planetary Science Letters},
+  keywords = {ionic radius,lattice strain,rare earth elements,trace elements},
+  number = {3}
+}
+
+@book{Deer2013,
+  title = {An Introduction to the Rock-Forming Minerals},
+  author = {Deer, W.A. and Howie, R.A. and Zussman, J.},
+  year = {2013},
+  edition = {3},
+  publisher = {{Mineralogical Society of Great Britain and Ireland}},
+  address = {{Twickenham, UK}},
+  isbn = {978-0-903056-27-4}
+}
+
+@article{Hunter2007,
+  title = {Matplotlib: {{A 2D Graphics Environment}}},
+  shorttitle = {Matplotlib},
+  author = {Hunter, John D.},
+  year = {2007},
+  volume = {9},
+  pages = {90--95},
+  issn = {1521-9615},
+  doi = {10.1109/MCSE.2007.55},
+  journal = {Computing in Science \& Engineering},
+  number = {3}
+}
+
+@article{LeBas1992,
+  title = {The Construction of the {{Total Alkali}}-{{Silica}} Chemical Classification of Volcanic Rocks},
+  author = {Le Bas, M. J. and Le Maitre, R. W. and Woolley, A. R.},
+  year = {1992},
+  month = mar,
+  volume = {46},
+  pages = {1--22},
+  issn = {0930-0708, 1438-1168},
+  doi = {10.1007/BF01160698},
[…]
