+++ paper/paper.md
@@ -1,0 +1,45 @@
+---
+title: 'pyrolite: Python for geochemistry'
+tags:
+  - Python
+  - geochemistry
+  - geology
+  - mineralogy
+  - petrology
+authors: # (Multiple affiliations must be quoted)
+  - name: Morgan J. Williams
+    orcid: 0000-0003-4764-9555
+    affiliation: "1"
+  - name: Louise Schoneveld
+    orcid: 0000-0002-9324-1676
+    affiliation: "1"
+  - name: Yajing Mao
+    orcid: 0000-0002-2725-2158
+    affiliation: "2"
+  - name: Jens Klump
+    orcid: 0000-0001-5911-6022
+    affiliation : "1"
+  - name: Justin Gosses
+    orcid: 0000-0002-5351-7295
+    affiliation: "3"
+  - name: Hayden Dalton
+    orcid: 0000-0003-2114-9894
+    affiliation: "4"
+  - name: Adam Bath
+    orcid: 0000-0003-0882-0807
+    affiliation: "1"
+  - name: Steve Barnes
+    orcid: 0000-0002-4912-9177
+    affiliation: "1"
+affiliations:
+ - name: CSIRO Mineral Resources
+   index: 1
+ - name: Institute of Geology and Geophysics, Chinese Academy of Geosciences
+   index: 2
+ - name: SAIC
+   index: 3
+ - name: School of Earth Sciences, University of Melbourne
+   index: 4
+date: 14 January 2020
+bibliography: paper.bib
+---
