### 2020-01-14 768e7e8
Add Aitchison1984 to paper.bib
