+++ paper/paper.bib
@@ -1,3 +1,16 @@
+
+@article{Aitchison1984,
+  title = {The Statistical Analysis of Geochemical Compositions},
+  author = {Aitchison, John},
+  year = {1984},
+  volume = {16},
+  pages = {531--564},
+  doi = {10.1007/BF01029316},
+  journal = {Journal of the International Association for Mathematical Geology},
+  keywords = {compositional data},
+  number = {6}
+}
+
 @article{Blundy2003,
   ids = {Blundy2003a},
   title = {Partitioning of Trace Elements between Crystals and Melts},
