### 2020-01-14 717e2de
Update paper.md
