### 2020-01-14 f1781f9
Add Li and Ripley 2009 to paper.bib
