+++ paper/paper.bib
@@ -79,6 +79,21 @@
   number = {5}
 }
 
+@article{Li2009,
+  title = {Sulfur {{Contents}} at {{Sulfide}}-{{Liquid}} or {{Anhydrite Saturation}} in {{Silicate Melts}}: {{Empirical Equations}} and {{Example Applications}}},
+  shorttitle = {Sulfur {{Contents}} at {{Sulfide}}-{{Liquid}} or {{Anhydrite Saturation}} in {{Silicate Melts}}},
+  author = {Li, Chusi and Ripley, Edward M.},
+  year = {2009},
+  month = may,
+  volume = {104},
+  pages = {405--412},
+  issn = {0361-0128},
+  doi = {10.2113/gsecongeo.104.3.405},
+  journal = {Economic Geology},
+  language = {en},
+  number = {3}
+}
+
 @inproceedings{McKinney2010,
   title = {Data Structures for Statistical Computing in Python},
   booktitle = {Proceedings of the 9th Python in Science Conference},
