### 2020-01-14 91cac75
Update paper.md text
