### 2020-01-14 183f28a
Add spider, heatscatter images to paper.md
