### 2020-01-15 #17 @lwasser
hi @morganjwilliams !! my apologies for the delay. i'm catching up post holidays and beginning of the semester. We have a meeting this thursday but i suspect we will ask for a full submission after that meeting. you are welcome to come to pitch the package as well if you'd like!! it's at 11am mountain time. If not i'll just bring it up as an issue.

In the meantime if you could think of people who would be good to review this package, that would be wonderful.
