### 2020-01-16 #17 @morganjwilliams
Thanks for the update, sounds good. I won't make the meeting this time around (2am my time), but I'll jot down some ideas for reviewers and start getting things ready for the full submission.
