### 2020-01-16 490325f
Minor updates for spider.spider
