### 2020-01-16 e23c75d
Update spider docs example
