### 2020-01-16 dbe04f9
Update spider example
