### 2020-01-17 f501b38
Update spider image for paper.md
