### 2020-01-20 1910b94
Update paper.bib with tidy data citation
