+++ paper/paper.bib
@@ -158,3 +158,20 @@
   language = {en},
   number = {2}
 }
+
+@article{Wickham2014,
+  title = {Tidy {{Data}}},
+  author = {Wickham, Hadley},
+  year = {2014},
+  month = sep,
+  volume = {59},
+  pages = {1--23},
+  issn = {1548-7660},
+  doi = {10.18637/jss.v059.i10},
+  copyright = {Copyright (c) 2013 Hadley  Wickham},
+  journal = {Journal of Statistical Software},
+  language = {en},
+  number = {1}
+}
+
+
