### 2020-01-20 3e46174
Update citations for paper.md
