+++ paper/paper.md
@@ -52,18 +52,18 @@ bibliography: paper.bib
 
 ## Features
 
-A variety of standard diagram methods (e.g. ternary, spider, and data-density diagrams), templated diagrams (e.g. the Total-Alkali Silica diagram [@LeBas1992] and Pearce diagrams [@Pearce2008]) and novel geochemical visualisation methods are available.
-
-![Example of different bivariate and ternary diagrams, highlighting the ability to visualise data distribution.](sphx_glr_heatscatter_001.png)
-
-![Example spider diagram, with comparison to a data-density based equivalent.](sphx_glr_spider_005.png)
-
+A variety of standard diagram methods (e.g. ternary, spider, and data-density diagrams), templated diagrams [e.g. the Total-Alkali Silica diagram , @LeBas1992; and Pearce diagrams, @Pearce2008] and novel geochemical visualisation methods are available.
 
 Reference datasets of compositional reservoirs (e.g. CI-Chondrite, Bulk Silicate Earth, Mid-Ocean Ridge Basalt) and a number of rock-forming mineral endmembers are installed with ``pyrolite``. The first of these enables normalisation of composition to investigate relative geochemical patterns, and the second facilitates mineral endmember recalculation and normative calculations.
 
 ``pyrolite`` also includes some specific methods to model geochemical patterns, such as the lattice strain model for trace element partitioning of @Blundy2003, the Sulfur Content at Sulfur Saturation (SCSS) model of @Li2009, and orthogonal polynomial decomposition for parameterising Rare Earth Element patterns of @ONeill2016.
 
 Extensions beyond the core functionality are also being developed, including ``pyrolite-meltsutil`` which provides utilities for working with ``alphaMELTS`` [@Smith2005] and it's outputs, and is targeted towards performing large numbers of related melting and fractionation experiments.
+
+![Example of different bivariate and ternary diagrams, highlighting the ability to visualise data distribution.](sphx_glr_heatscatter_001.png)
+
+![Example spider diagram, with comparison to a data-density based equivalent.](sphx_glr_spider_005.png)
+
 
 ## Conventions
 
@@ -72,7 +72,7 @@ Extensions beyond the core functionality are also being developed, including ``p
 Tidy Geochemical Tables
 </dt>
 
-Being based on ``pandas``, ``pyrolite`` operations are based on tabular structured data in dataframes, where each geochemical variable or component is a column, and each observation is a row (consistent with 'tidy data' principles, @Wickham2014).  ``pyrolite`` additionally assumes that geochemical components are identifiable with either element- or oxide-based column names (which contain only one element excluding oxygen, e.g. Ca, MgO, Al<sub>2</sub>O<sub>3</sub>,
+Being based on ``pandas``, ``pyrolite`` operations are based on tabular structured data in dataframes, where each geochemical variable or component is a column, and each observation is a row [consistent with 'tidy data' principles, @Wickham2014]. ``pyrolite`` additionally assumes that geochemical components are identifiable with either element- or oxide-based column names (which contain only one element excluding oxygen, e.g. Ca, MgO, Al<sub>2</sub>O<sub>3</sub>,
 but not Ca<sub>3</sub>Al<sub>3</sub>(SiO<sub>4</sub>)<sub>3</sub> or Ca_Wt%).
 
 <dt>
