### 2020-01-20 3c52077
Update element/formula formatting
