### 2020-01-20 3af17fa
Update Fig 1 position
