### 2020-01-21 0673715
Update density example, relim and logscale bug
