### 2020-01-21 98ce4bd
Update stem example
Update stem example
