### 2020-01-22 02900c8
Add link through to docs installation page for Readme
