### 2020-01-22 cb9f16f
Add getting started guide note
