### 2020-01-22 eb3f4a2
Update features section for paper.md
