### 2020-01-22 5da884d
Add Hofmann 2014 to bibliography
