+++ paper/paper.bib
@@ -37,6 +37,23 @@
   isbn = {978-0-903056-27-4}
 }
 
+@incollection{Hofmann2014,
+  title = {3.3 - {{Sampling Mantle Heterogeneity}} through {{Oceanic Basalts}}: {{Isotopes}} and {{Trace Elements}}},
+  shorttitle = {3.3 - {{Sampling Mantle Heterogeneity}} through {{Oceanic Basalts}}},
+  booktitle = {Treatise on {{Geochemistry}} ({{Second Edition}})},
+  author = {Hofmann, A. W.},
+  editor = {Holland, Heinrich D. and Turekian, Karl K.},
+  year = {2014},
+  month = jan,
+  pages = {67--101},
+  publisher = {{Elsevier}},
+  address = {{Oxford}},
+  doi = {10.1016/B978-0-08-095975-7.00203-5},
+  abstract = {This chapter is in part an update of a previous, more abbreviated review (Hofmann, 1997). It covers the subject in greater depth, and it reflects some significant changes in the author's views since the writing of the earlier paper. In particular, the spatial range of equilibrium attained during partial melting may be much smaller than previously thought, because of new experimental diffusion data and new results from natural settings. Also, the question of `layered' versus `whole-mantle' convection, including the depth of subduction and of the origin of plumes, has to be reassessed in light of the recent breakthroughs achieved by seismic mantle tomography. As the spatial resolution of seismic tomography and the pressure range, accuracy, and precision of experimental data on melting relations, phase transformations, and kinetics continue to improve, the interaction between these disciplines and geochemistry sensu stricto will continue to improve our understanding of what is actually going on in the mantle. The established views of the mantle being engaged in simple two- or single-layer convection are becoming obsolete. In many ways, we are just at the beginning of this new phase of mantle geology, geophysics, and geochemistry.},
+  isbn = {978-0-08-098300-4},
+  keywords = {Basalts’ lead paradox,Crust-Mantle differentiation,Local equilibrium,Major elements,Mantle geochemistry,Regional equilibrium,Trace elements}
+}
+
 @article{Hunter2007,
   title = {Matplotlib: {{A 2D Graphics Environment}}},
   shorttitle = {Matplotlib},
