### 2020-01-22 2b805ee
Update paper.md community section
