### 2020-01-22 f4ef59b
Add references to paper.bib
