+++ paper/paper.bib
@@ -135,6 +135,19 @@
   number = {8}
 }
 
+@article{Palme2014,
+  title = {Cosmochemical {{Estimates}} of {{Mantle Composition}}},
+  author = {Palme, H. and O'Neill, H. St. C.},
+  editor = {Holland, Heinrich D. and Turekian, Karl K.},
+  year = {2014},
+  volume = {3},
+  pages = {1--39},
+  doi = {10.1016/B978-0-08-095975-7.00201-1},
+  abstract = {The composition of the primitive mantle derived here shows that Earth was assembled from material that shows many of the same chemical fractionation processes as chondritic meteorites. These processes occurred at the initial stage of the solar system formation, under conditions thought to be present in the solar nebula. But the stable isotope record excludes chondritic meteorites as the `building blocks' of Earth. Meteorites formed in local environments separated from that part of the inner solar system where much of the material forming the terrestrial planets was sourced.},
+  journal = {Treatise on Geochemistry (Second Edition)},
+  keywords = {Chondritic meteorites,Collisional erosion,Depletion of volatile elements,Earth's bulk composition,Earth's primitive mantle,Late veneer,Stable isotopes in Earth and meteorites}
+}
+
 @article{Pearce2008,
   title = {Geochemical Fingerprinting of Oceanic Basalts with Applications to Ophiolite Classification and the Search for {{Archean}} Oceanic Crust},
   author = {Pearce, Julian A.},
@@ -174,6 +187,25 @@
   keywords = {geochemical software,mantle melting,MELTS,pHMELTS,pMELTS,trace elements},
   language = {en},
   number = {2}
+}
+
+@article{Sun1989,
+  title = {Chemical and Isotopic Systematics of Oceanic Basalts: Implications for Mantle Composition and Processes},
+  shorttitle = {Chemical and Isotopic Systematics of Oceanic Basalts},
+  author = {Sun, S.-s and McDonough, W. F.},
+  year = {1989},
+  month = jan,
+  volume = {42},
+  pages = {313--345},
+  issn = {0305-8719, 2041-4927},
+  doi = {10.1144/GSL.SP.1989.042.01.19},
+  abstract = {Skip to Next Section
+Trace-element data for mid-ocean ridge basalts (MORBs) and ocean island basalts (OIB) are used to formulate chemical systematics for oceanic basalts. The data suggest that the order of trace-element incompatibility in oceanic basalts is Cs {$\approx$} Rb {$\approx$} ({$\approx$} Tl) {$\approx$} Ba({$\approx$} W) {$>$} Th {$>$} U {$\approx$} Nb = Ta {$\approx$} K {$>$} La {$>$} Ce {$\approx$} Pb {$>$} Pr ({$\approx$} Mo) {$\approx$} Sr {$>$} P {$\approx$} Nd ({$>$} F) {$>$} Zr = Hf {$\approx$} Sm {$>$} Eu {$\approx$} Sn ({$\approx$} Sb) {$\approx$} Ti {$>$} Dy {$\approx$} (Li) {$>$} Ho = Y {$>$} Yb. This rule works in general and suggests that the overall fractionation processes operating during magma generation and evolution are relatively simple, involving no significant change in the environment of formation for MORBs and OIBs.
[…]
