### 2020-01-22 ffb0290
Add Pearson citation to paper.bib
