+++ paper/paper.bib
@@ -160,6 +160,19 @@
   number = {1}
 }
 
+@article{Pearson1897,
+  title = {Mathematical Contributions to the Theory of Evolution.\textemdash{{On}} a Form of Spurious Correlation Which May Arise When Indices Are Used in the Measurement of Organs},
+  author = {Pearson, Karl},
+  year = {1897},
+  month = jan,
+  volume = {60},
+  pages = {489--498},
+  doi = {10.1098/rspl.1896.0076},
+  abstract = {(1) If the ratio of two absolute measurements on the same or different organs be taken it is convenient to term this ratio an index. If u = f1(x, y) and v = f2(x, y) be two functions of the three variables x, y, z, and these variables be selected at random so that there exists no correlation between x, y, y, z, or z, x, there will still be found to exist correlation between u and v. Thus a real danger arises when a statistical biologist attribntes the correlation between two functions like u and v to organic relationship. The particular case that is likely to occur is when u and v are indices with the same denominator for the correlation of indices seems at first sight a very plausible measure of organic correlation.},
+  journal = {Proceedings of the Royal Society of London},
+  number = {359-367}
+}
+
 @article{Pedregosa2011,
   title = {Scikit-Learn: {{Machine Learning}} in {{Python}}},
   shorttitle = {Scikit-Learn},
