### 2020-01-23 653d5f2
HD Edits for paper.md
