+++ paper/paper.md
@@ -54,7 +54,7 @@ The project aims to contribute to more robust, efficient and reproducible data-d
 
 ``pyrolite`` provides tools for processing, transforming and visualising geochemical data from common tabular formats.
 The package includes methods to recalculate and rescale whole-rock and mineral compositions, perform compositional statistics and create appropriate visualisations and also includes numerous auxiliary utilities (e.g. a geological timescale).
-These tools also provide a foundation for preparing data for subsequent machine learning applications using ``scikit-learn``  [@Pedregosa2011].
+In addition, these tools provide a foundation for preparing data for subsequent machine learning applications using ``scikit-learn``  [@Pedregosa2011].
 
 Geochemical data are compositional (i.e. sum to 100%), and as such require non-standard statistical treatment [@Aitchison1984]. While challenges of compositional data have long been acknowledged [e.g. @Pearson1897], appropriate measures to account for this have thus far seen limited uptake by the geochemistry community. The submodule ``pyrolite.comp`` provides access to methods for transforming compositional data, facilitating more robust statistical practises.
 
@@ -104,7 +104,7 @@ This convention is equivalent to assuming that the system is open to oxygen, and
 # Community
 
 ``pyrolite`` aims to be designed, developed and supported by the geochemistry community.
-Community contributions are encouraged, and will help make ``pyrolite`` a broadly useful toolkit and resource (especially for e.g. teaching purposes).
+Community contributions are encouraged, and will help make ``pyrolite`` a broadly useful toolkit and resource (for both research and education purposes).
 In addition to developing a library of commonly used methods and diagram templates, these contributions will contribute to enabling better research practices, and potentially even establishing standards for geochemical data processing and analysis within the user community.
 
 # Acknowledgements
