### 2020-01-23 7d61850
Minor text formatting
