### 2020-01-29 29d472d
Disable automatic extension loading
