### 2020-01-29 f72d3ce
Move check_perl to pyrolite_meltsutil
