### 2020-02-10 1343b1a
Take EMCOMP draft out of docs
Remove EMCOMP built files
