### 2020-02-10 #17 @morganjwilliams
Thanks @lwasser, I'm getting onto this now. It'll go into our internal publications system today, and should be able to be submitted shortly afterwards. I have an idea of one or two reviewers who might be interested, I'll ping them in the issue thread when I get the submission in.
