### 2020-02-14 3de9c30
Add pd.Series support for color indexing
Update pd.Series support for color arg
