### 2020-02-14 cb9b6e3
Workaround for using ax reference after converting to ternary
