### 2020-02-17 e1c7360
Add partial support for categorical coloring
