### 2020-02-17 99a1d8c
Minor update for geochem.magma.scss
Bugfix geochem.magma.scss
