+++ docs/source/dev/changelog.rst
@@ -7,8 +7,6 @@ All notable changes to this project will be documented here.
 Todo
 ~~~~~~~
 
-* FEATURE: Incompatible element indexes for :mod:`pyrolite.geochem`
-  (`#30 <https://github.com/morganjwilliams/pyrolite/issues/30>`__)
 * FEATURE: Updates to include more lithogeochemical plot templates
   (`#26 <https://github.com/morganjwilliams/pyrolite/issues/26>`__)
 * BUG: Upgrades for :func:`~pyrolite.geochem.transform.convert_chemistry` for performance,
@@ -26,10 +24,42 @@ Todo
         If you're keen to check something out before its released, you can use a
         `development install <development.html#development-installation>`__.
 
+* Added an 'Importing Data' section to the docs
+  `Getting Started page <../gettingstarted.html#Importing-Data>`__.
+* Disabled automatic extension loading (e.g. for :mod:`pyrolite_meltsutil`) to
+  avoid bugs during version mismatches.
+* Moved :func:`~pyrolite.util.general.check_perl` out of :mod:`pyrolite` into
+  :mod:`pyrolite_meltsutil`
+* PR Merged: `@lavender22 <https://github.com/lavender22>`__ updated the spider
+  diagram example to add a link to the normalisation example (which lists
+  different reservoirs you can normalise to).
+
+:mod:`pyrolite.plot`
+~~~~~~~~~~~~~~~~~~~~~~~~~
+
+* Added support for spider plot index ordering added with the keyword
+  :code:`index_order` (`#30 <https://github.com/morganjwilliams/pyrolite/issues/30>`__)
+* Added support for color indexing in :mod:`~pyrolite.plot.color` using
+  :class:`pandas.Series`, and also for list-like arrays of categories
+* Added a workaround for referring to axes positions where the projection is changed
+  to a ternary projection (displacing the original axis), but the reference to the
+  original axes object (now booted from :code:`fig.axes`/:code:`fig.orderedaxes`) is
+  subsequently used. 
+* Updated :func:`~pyrolite.plot.color.process_color` processing of auxillary
+  color keyword arguments (fixing a bug for color arguments in
+  :func:`~pyrolite.plot.stem`)
 * Updated the effect of :code:`relim` keyword argument of
   :func:`~pyrolite.plot.density.density` to remove the scaling (it will no longer
   log-scale the axes, just the grid/histogram bins).
 
+:mod:`pyrolite.geochem`
+~~~~~~~~~~~~~~~~~~~~~~~~~
+* Added sorting function :code:`~pyrolite.geochem.ind.by_imcompatibility`
+  for incompatible element sorting (based on BCC/PM relative abundances).
+
+:mod:`pyrolite.mineral`
+~~~~~~~~~~~~~~~~~~~~~~~~~
+* Minor bugfix for :func:`~pyrolite.mineral.mindb.update_database`
 
 `0.2.4`_
 --------------
