### 2020-02-17 7b00010
Minor docs updates
