### 2020-02-18 0dd84e3
Updates to pyrolite.pyrocomp API
