+++ pyrolite/comp/__init__.py
@@ -45,6 +45,249 @@ class pyrocomp(object):
         --------
         :class:`pandas.DataFrame`
             Renormalized dataframe.
+
+        Notes
+        ------
+        This won't modify the dataframe in place, you'll need to assign it to something.
+        If you specify components, those components will be summed to 100%,
+        and others remain unchanged.
         """
         obj = self._obj
         return renormalise(obj, components=components, scale=scale)
+
+    def ALR(self, components=[], ind=-1, null_col=False):
+        """
+        Additive Log Ratio transformation.
+
+        Parameters
+        ---------------
+        ind: :class:`int`, :class:`str`
+            Index or name of column used as denominator.
+        null_col : :class:`bool`
+            Whether to keep the redundant column.
+
+        Returns
+        ---------
+        :class:`pandas.DataFrame`
+            ALR-transformed array, of shape :code:`(N, D-1)`.
+        """
+        components = self._obj.columns.values.tolist()
+
+        if isinstance(ind, int):
+            index_col_no = ind
+        elif isinstance(ind, str):
+            assert ind in components
+            index_col_no = components.index(ind)
+        if index_col_no == -1:
+            index_col_no += len(components)
+        index_col = components[index_col_no]
+        colnames = ["ALR({}/{})".format(c, index_col) for c in components]
+
+        if not null_col:
+            colnames = [n for ix, n in enumerate(colnames) if ix != index_col_no]
+        tfm_df = pd.DataFrame(
+            alr(self._obj[components].values, ind=index_col_no, null_col=null_col),
+            index=self._obj.index,
+            columns=colnames,
+        )
+        tfm_df.alr_index = index_col_no  # save parameter for inverse_transform
+        tfm_df.inverts_to = self._obj.columns.to_list()
+        return tfm_df
+
+    def inverse_ALR(self, ind=None, null_col=False):
+        """
+        Inverse Additive Log Ratio transformation.
+
+        Parameters
+        ---------------
+        ind: :class:`int`, :class:`str`
+            Index or name of column used as denominator.
+        null_col : :class:`bool`, :code:`False`
+            Whether the array contains an extra redundant column
+            (i.e. shape is :code:`(N, D)`).
+
+        Returns
+        --------
+        :class:`pandas.DataFrame`
+            Inverse-ALR transformed array, of shape :code:`(N, D)`.
+        """
+
+        if hasattr(self._obj, "inverts_to"):
+            colnames = self._obj.inverts_to
+
+        if ind is None:
+            ind = self._obj.alr_index or -1
+
+        itfm_df = pd.DataFrame(
+            inverse_alr(self._obj.values, ind=ind, null_col=null_col),
+            index=self._obj.index,
+            columns=colnames,
+        )
+        return itfm_df
+
+    def CLR(self):
+        """
+        Centred Log Ratio transformation.
+
+        Parameters
+        ---------------
+
+        Returns
+        ---------
+        :class:`pandas.DataFrame`
+            CLR-transformed array, of shape :code:`(N, D)`.
+        """
+        colnames = ["CLR({}/g)".format(c) for c in self._obj.columns]
+        tfm_df = pd.DataFrame(
+            clr(self._obj.values), index=self._obj.index, columns=colnames,
+        )
+        tfm_df.inverts_to = (
+            self._obj.columns.to_list()
+        )  # save parameter for inverse_transform
+        return tfm_df
+
+    def inverse_CLR(self):
+        """
+        Inverse Centred Log Ratio transformation.
+
+        Parameters
+        --------------
+
+        Returns
+        ---------
+        :class:`pandas.DataFrame`
[…]
