### 2020-02-18 df4a321
Update color mapping support within plot.color
