### 2020-02-18 0abec8a
Update plot.color.process_color docstring
