### 2020-02-19 d6dd377
Update plot.ternary to use grid edges for pcolormesh
This fixes a bug where shading='gouraud' wouldn't propagate to pcolormesh due to mismatched array sizes
