### 2020-02-19 7164c30
Update plot.ternary.grid to enable separate x and y bins
This enables grid bins to be specified with either bins=x or bins = (x, y); minimizing computation cost for evaluating the KDE
