+++ pyrolite/plot/density/grid.py
@@ -12,10 +12,34 @@ class DensityGrid(object):
     def __init__(
         self, x, y, extent=None, bins=50, logx=False, logy=False, coverage_scale=1.2
     ):
+        """
+        Build a grid of x-y coordinates for use in evaluating KDE functions.
+
+        Parameters
+        -----------
+        x : :class:`np.ndarray`
+        y : :class:`np.ndarray`
+        extent : :class:`list`
+            Optionally-specified extent for the grid in the form (xmin, xmax, ymin, ymax).
+        bins : :class:`int` | :class:`tuple`
+            Number of bins for the grid. Can optionally specify
+            a tuple with (xbins, ybins).
+        logx : :class:`bool`
+            Whether to use a log-spaced index for the x dimension of the grid.
+        logy: :class:`bool`
+            Whether to use a log-spaced index for the y dimension of the grid.
+        coverage_scale : :class:`float`
+            Multiplier for the range of the grid relative to the data. If >1, the grid
+            will extend beyond the data.
+        """
         # limits
         self.logx = logx
         self.logy = logy
-        self.bins = bins
+        if not isinstance(bins, int):
+            assert len(bins) == 2  # x-y bins
+            self.xbins, self.ybins = bins
+        else:
+            self.xbins, self.ybins = bins, bins
         self.coverage_scale = coverage_scale
 
         if extent is None:
@@ -41,15 +65,15 @@ class DensityGrid(object):
 
     def get_ystep(self):
         if self.logy:
-            return (self.ymax / self.ymin) / self.bins
+            return (self.ymax / self.ymin) / self.ybins
         else:
-            return (self.ymax - self.ymin) / self.bins
+            return (self.ymax - self.ymin) / self.ybins
 
     def get_xstep(self):
         if self.logx:
-            return (self.xmax / self.xmin) / self.bins
+            return (self.xmax / self.xmin) / self.xbins
         else:
-            return (self.xmax - self.xmin) / self.bins
+            return (self.xmax - self.xmin) / self.xbins
 
     def extent_from_xy(self, x, y, coverage_scale=None):
         cov = coverage_scale or self.coverage_scale
@@ -73,14 +97,14 @@ class DensityGrid(object):
 
     def update_grid_centre_ticks(self):
         if self.logx:
-            self.grid_xc = logspc_(self.xmin, self.xmax, 1.0, self.bins)
+            self.grid_xc = logspc_(self.xmin, self.xmax, 1.0, self.xbins)
         else:
-            self.grid_xc = linspc_(self.xmin, self.xmax, self.xstep, self.bins)
+            self.grid_xc = linspc_(self.xmin, self.xmax, self.xstep, self.xbins)
 
         if self.logy:
-            self.grid_yc = logspc_(self.ymin, self.ymax, 1.0, self.bins)
+            self.grid_yc = logspc_(self.ymin, self.ymax, 1.0, self.ybins)
         else:
-            self.grid_yc = linspc_(self.ymin, self.ymax, self.ystep, self.bins)
+            self.grid_yc = linspc_(self.ymin, self.ymax, self.ystep, self.ybins)
 
     def update_grid_edge_ticks(self):
         self.update_grid_centre_ticks()
@@ -114,18 +138,35 @@ class DensityGrid(object):
             yex = [self.ymin - self.ystep, self.ymax + self.ystep]
         return xex + yex
 
-    def kdefrom(self, xy, xtransform=lambda x: x, ytransform=lambda x: x):
-        assert np.isfinite(self.grid_xc).all() and np.isfinite(self.grid_yc).all()
+    def kdefrom(
+        self, xy, xtransform=lambda x: x, ytransform=lambda x: x, mode="centres"
+    ):
+        """
[…]
