### 2020-02-19 faf5986
Update typo for paper.md
