### 2020-02-19 1680464
Fix changelog typo
