### 2020-02-20 1ed3c83
Add codacy.yml
