### 2020-02-20 3f02e9a
Update codacy.yml to exclude github folder
