### 2020-02-20 883f266
Clean up util.skl.select
