+++ pyrolite/util/skl/select.py
@@ -1,16 +1,15 @@
 import logging
-import numpy as np
 import pandas as pd
 from ...geochem.ind import __common_elements__, __common_oxides__, REE
+
+logging.getLogger(__name__).addHandler(logging.NullHandler())
+logger = logging.getLogger(__name__)
 
 try:
     from sklearn.base import TransformerMixin, BaseEstimator
 except ImportError:
     msg = "scikit-learn not installed"
     logger.warning(msg)
-
-logging.getLogger(__name__).addHandler(logging.NullHandler())
-logger = logging.getLogger(__name__)
 
 
 class TypeSelector(BaseEstimator, TransformerMixin):
@@ -27,6 +26,9 @@ class TypeSelector(BaseEstimator, TransformerMixin):
 
 class ColumnSelector(BaseEstimator, TransformerMixin):
     def __init__(self, columns):
+        """
+        Select specific columns from a dataframe for further transformation.
+        """
         self.columns = columns
 
     def fit(self, X, y=None):
@@ -45,9 +47,12 @@ class ColumnSelector(BaseEstimator, TransformerMixin):
 
 
 class CompositionalSelector(BaseEstimator, TransformerMixin):
-    def __init__(
-        self, components=__common_elements__ | __common_oxides__, inverse=False
-    ):
+    def __init__(self, components=None, inverse=False):
+        """
+        Select the oxide and element components from a dataframe.
+        """
+        if components is None:
+            components = __common_elements__ | __common_oxides__
         self.columns = components
         self.inverse = inverse
 
@@ -65,7 +70,12 @@ class CompositionalSelector(BaseEstimator, TransformerMixin):
 
 
 class MajorsSelector(BaseEstimator, TransformerMixin):
-    def __init__(self, components=__common_oxides__):
+    def __init__(self, components=None):
+        """
+        Select the major element oxides from a dataframe.
+        """
+        if components is None:
+            components = __common_oxides__
         self.columns = components
 
     def fit(self, X, y=None):
@@ -79,7 +89,12 @@ class MajorsSelector(BaseEstimator, TransformerMixin):
 
 
 class ElementSelector(BaseEstimator, TransformerMixin):
-    def __init__(self, components=__common_elements__):
+    def __init__(self, components=None):
+        """
+        Select the (trace) elements from a dataframe.
+        """
+        if components is None:
+            components = __common_elements__
         self.columns = components
 
     def fit(self, X, y=None):
@@ -93,7 +108,12 @@ class ElementSelector(BaseEstimator, TransformerMixin):
 
 
 class REESelector(BaseEstimator, TransformerMixin):
-    def __init__(self, components=REE()):
+    def __init__(self, components=None):
+        """
+        Select the Rare Earth Elements (REE) from a dataframe.
+        """
+        if components is None:
+            components = REE()
         components = [i for i in components if not i == "Pm"]
         self.columns = components
 
