### 2020-02-20 8ce7272
Update util.skl.select docstrings
