### 2020-02-20 e0c1e60
Add codacy coverage to travis
