+++ .travis.yml
@@ -11,6 +11,10 @@ jobs:
     - python: 3.8
       env: PYVERSION=38
 
+before_install:
+  - sudo apt-get install jq
+  - curl -LSs $(curl -LSs https://api.github.com/repos/codacy/codacy-coverage-reporter/releases/latest | jq -r '.assets | map({name, browser_download_url} | select(.name | endswith(".jar"))) | .[0].browser_download_url')" -o codacy-coverage-reporter-assembly.jar
+
 install:
   - pip install -q pytest pytest-runner pytest-cov coverage
   - pip install coveralls
@@ -21,4 +25,5 @@ script:
 
 after_success:
   - "ls -lr htmlcov"
-  - coveralls
+  - coveralls # coverage to coveralls.io
+  - java -jar codacy-coverage-reporter-assembly.jar report -l Java -r build/reports/jacoco/test/jacocoTestReport.xml # coverage to codacy.com
