### 2020-02-20 a5f14ad
util.skl.select - tests and docstrings
