### 2020-02-20 be2f56a
Update codacy coverage
