### 2020-02-20 5960462
Add codacy-coverage to travis config
