+++ setup.cfg
@@ -2,7 +2,7 @@
 test=pytest
 
 [tool:pytest]
-addopts = test --cov=pyrolite --cov-report html --cov-report term-missing --cov-config .coveragerc
+addopts = test --cov=pyrolite --cov-report html --cov-report xml --cov-report term-missing --cov-config .coveragerc
 python_files = *.py
 
 [versioneer]
