### 2020-02-20 d68fde0
Add Codacy badge
