+++ pyrolite/util/wfs.py
@@ -1,10 +1,0 @@
-from owslib.wfs import WebFeatureService
-from owslib.fes import *
-from owslib.etree import etree
-import requests
-import json, geojson
-import codecs
-import logging
-
-logging.getLogger(__name__).addHandler(logging.NullHandler())
-logger = logging.getLogger(__name__)
