+++ pyrolite/util/database.py
@@ -1,6 +1,7 @@
 import sys
 import struct
 from contextlib import contextmanager
+from tinydb import TinyDB
 
 __backend__ = None
 try:
