### 2020-02-20 2ac455e
Cleanup util.general
