### 2020-02-20 006730b
Clean up util.classification
