### 2020-02-20 4d52d56
Updates for mineral to reduce duplication
