### 2020-02-20 2e27b4d
Cleanup util.plot
Minor reformatting
