### 2020-02-20 937c09f
Clean up plot.density
