### 2020-02-20 8c9e6b4
Add mpltern import to make ternary projection available
