### 2020-02-24 041ecbc
Modify setup.py note
