### 2020-02-24 93a8cef
Clean up comp.__init__
