+++ pyrolite/comp/__init__.py
@@ -8,7 +8,17 @@ import logging
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 logger = logging.getLogger(__name__)
 
-from .codata import *
+from .codata import (
+    renormalise,
+    alr,
+    clr,
+    ilr,
+    boxcox,
+    inverse_alr,
+    inverse_clr,
+    inverse_ilr,
+    inverse_boxcox,
+)
 from .aggregate import *
 from .impute import *
 
@@ -16,11 +26,10 @@ from .impute import *
 @pd.api.extensions.register_series_accessor("pyrocomp")
 @pd.api.extensions.register_dataframe_accessor("pyrocomp")
 class pyrocomp(object):
-    """
-    Custom dataframe accessor for pyrolite compositional transforms.
-    """
-
     def __init__(self, obj):
+        """
+        Custom dataframe accessor for pyrolite compositional transforms.
+        """
         self._validate(obj)
         self._obj = obj
 
@@ -33,7 +42,7 @@ class pyrocomp(object):
         Renormalises compositional data to ensure closure.
 
         Parameters
-        ------------
+        ----------
         components : :class:`list`
             Option subcompositon to renormalise to 100. Useful for the use case
             where compostional data and non-compositional data are stored in the
@@ -42,7 +51,7 @@ class pyrocomp(object):
             Closure parameter. Typically either 100 or 1.
 
         Returns
-        --------
+        -------
         :class:`pandas.DataFrame`
             Renormalized dataframe.
 
@@ -60,14 +69,14 @@ class pyrocomp(object):
         Additive Log Ratio transformation.
 
         Parameters
-        ---------------
+        ----------
         ind: :class:`int`, :class:`str`
             Index or name of column used as denominator.
         null_col : :class:`bool`
             Whether to keep the redundant column.
 
         Returns
-        ---------
+        -------
         :class:`pandas.DataFrame`
             ALR-transformed array, of shape :code:`(N, D-1)`.
         """
@@ -99,7 +108,7 @@ class pyrocomp(object):
         Inverse Additive Log Ratio transformation.
 
         Parameters
-        ---------------
+        ----------
         ind: :class:`int`, :class:`str`
             Index or name of column used as denominator.
         null_col : :class:`bool`, :code:`False`
@@ -107,7 +116,7 @@ class pyrocomp(object):
             (i.e. shape is :code:`(N, D)`).
 
         Returns
-        --------
+        -------
         :class:`pandas.DataFrame`
             Inverse-ALR transformed array, of shape :code:`(N, D)`.
         """
@@ -130,10 +139,10 @@ class pyrocomp(object):
         Centred Log Ratio transformation.
 
         Parameters
-        ---------------
+        ----------
 
         Returns
-        ---------
+        -------
         :class:`pandas.DataFrame`
             CLR-transformed array, of shape :code:`(N, D)`.
         """
@@ -151,10 +160,10 @@ class pyrocomp(object):
         Inverse Centred Log Ratio transformation.
 
         Parameters
-        --------------
+        ----------
 
         Returns
-        ---------
+        -------
         :class:`pandas.DataFrame`
             Inverse-CLR transformed array, of shape :code:`(N, D)`.
         """
@@ -171,10 +180,10 @@ class pyrocomp(object):
         Isometric Log Ratio transformation.
 
         Parameters
-        ---------------
+        ----------
 
         Returns
-        --------
+        -------
         :class:`pandas.DataFrame`
             ILR-transformed array, of shape :code:`(N, D-1)`.
         """
@@ -192,7 +201,7 @@ class pyrocomp(object):
         Inverse Isometric Log Ratio transformation.
 
         Parameters
-        ---------------
+        ----------
         X : :class:`numpy.ndarray`, :code:`None`
[…]
