### 2020-02-24 68b0e46
Update comp.__init__ imports
