+++ pyrolite/comp/aggregate.py
@@ -1,4 +1,5 @@
 import numpy as np
+import pandas as pd
 import warnings
 from .codata import alr, inverse_alr
 
