### 2020-02-24 de42d99
Clean up to pyrolite.geochem.transform
