### 2020-02-24 c5fa01a
Cleanup geochem.norm
