### 2020-02-24 f9dfa0a
Clean up plot.__init__
