### 2020-02-24 9797607
Clean up util.time
