### 2020-02-24 9f88f05
Clean up util.skl.vis
