### 2020-02-24 255c3f7
Minor docstring updates for plot.templates.components
