### 2020-02-24 f6767fe
Minor cleanup for plot.spider
Spider cleanup
