### 2020-02-24 4673157
Minor cleanup for util.spatial
Fix missing np.
