+++ pyrolite/util/spatial.py
@@ -2,10 +2,8 @@
 Baisc spatial utility functions.
 """
 import numpy as np
-from numpy import cos, sin, deg2rad, arctan, arctan2, arcsin, arccos, sqrt, abs
 import itertools
 import logging
-from pyrolite.util.math import isclose
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 logger = logging.getLogger(__name__)
@@ -23,9 +21,9 @@ def _spherical_law_cosinse_GC_distance(ps):
     """
     φ1, φ2 = ps[2:]  # latitude
     λ1, λ2 = ps[:2]  # longitude
-    Δλ = abs(λ1 - λ2)
-    Δφ = abs(φ1 - φ2)
-    return arccos(sin(φ1) * sin(φ2) + cos(φ1) * cos(φ2) * cos(Δλ))
+    Δλ = np.abs(λ1 - λ2)
+    Δφ = np.abs(φ1 - φ2)
+    return np.arccos(np.sin(φ1) * np.sin(φ2) + np.cos(φ1) * np.cos(φ2) * np.cos(Δλ))
 
 
 def _vicenty_GC_distance(ps):
@@ -43,15 +41,15 @@ def _vicenty_GC_distance(ps):
     """
     φ1, φ2 = ps[2:]  # latitude
     λ1, λ2 = ps[:2]  # longitude
-    Δλ = abs(λ1 - λ2)
-    Δφ = abs(φ1 - φ2)
+    Δλ = np.abs(λ1 - λ2)
+    # Δφ = np.abs(φ1 - φ2)
 
-    _S = sqrt(
-        (cos(φ2) * sin(Δλ)) ** 2
-        + (cos(φ1) * sin(φ2) - sin(φ1) * cos(φ2) * cos(Δλ)) ** 2
+    _S = np.sqrt(
+        (np.cos(φ2) * np.sin(Δλ)) ** 2
+        + (np.cos(φ1) * np.sin(φ2) - np.sin(φ1) * np.cos(φ2) * np.cos(Δλ)) ** 2
     )
-    _C = sin(φ1) * sin(φ2) + cos(φ1) * cos(φ2) * cos(Δλ)
-    return np.abs(arctan2(_S, _C))
+    _C = np.sin(φ1) * np.sin(φ2) + np.cos(φ1) * np.cos(φ2) * np.cos(Δλ)
+    return np.abs(np.arctan2(_S, _C))
 
 
 def _haversine_GC_distance(ps):
@@ -62,13 +60,16 @@ def _haversine_GC_distance(ps):
     Parameters
     ----------
     ps : :class:`numpy.ndarray`
-        Numpy array wih latitudes and longitudes [x1, x2, y1, y2]
+        Numpy array wih latitudes and longitudes [x1, x2, y1, y2].
+
     """
     φ1, φ2 = ps[2:]  # latitude
     λ1, λ2 = ps[:2]  # longitude
-    Δλ = abs(λ1 - λ2)
-    Δφ = abs(φ1 - φ2)
-    return 2 * arcsin(sqrt(sin(Δφ / 2) ** 2 + cos(φ1) * cos(φ2) * sin(Δλ / 2) ** 2))
+    Δλ = np.abs(λ1 - λ2)
+    Δφ = np.abs(φ1 - φ2)
+    return 2 * np.arcsin(
+        np.sqrt(np.sin(Δφ / 2) ** 2 + np.cos(φ1) * np.cos(φ2) * np.sin(Δλ / 2) ** 2)
+    )
 
 
 def great_circle_distance(
@@ -91,13 +92,14 @@ def great_circle_distance(
     method : :class:`str`, :code:`{'vicenty', 'cosines', 'haversine'}`
[…]
