### 2020-02-24 7512516
Update codacy config
