+++ .codacy.yml
@@ -7,3 +7,5 @@ exclude_paths:
   - 'pyrolite/_version.py'
   - 'test/**'
   - '.github/**'
+  - 'CODE_OF_CONDUCT.md'
+  - 'CHANGELOG.md'
