### 2020-02-24 ca15282
Minor cleanup for plot.templates.TAS
