### 2020-02-24 436b1af
Comment out unused variable from util.spatial distance function
