### 2020-02-24 93a5ac8
Update limit defn for plot.templates.TAS
