### 2020-02-24 d10c148
Modify exception handling for util.classificaiton
