### 2020-02-24 ebca48a
Minor updates for geochem.ind and util
