+++ pyrolite/util/types.py
@@ -15,6 +15,7 @@ def iscollection(obj):
     -------
     :class:`bool`
         Boolean indication of whether the object is a collection.
+        
     """
 
     for ty in [list, np.ndarray, set, tuple, dict, pd.Series]:
