+++ pyrolite/util/text.py
@@ -5,7 +5,7 @@ import logging
 
 try:
     from sortedcollections import SortedSet as set
-except:
+except ImportError:
     pass
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
@@ -230,6 +230,7 @@ def slugify(value, delim="-"):
     Returns
     -------
     :class:`str`
+    
     """
     value = re.sub("[^\w\s-]", "", value).strip()
     value = re.sub("[-\s]+", delim, value)
@@ -242,14 +243,15 @@ def int_to_alpha(num):
     axes/figures.
 
     Parameters
-    -----------
+    ----------
     int : :class:`int`
         Integer to encode.
 
     Returns
-    ---------
+    -------
     :class:`str`
         Alpha-encoding of a small integer.
+
     """
     alphas = [chr(i).lower() for i in range(65, 65 + 26)]
     return alphas[num]
