+++ pyrolite/util/pd.py
@@ -1,10 +1,8 @@
 import pandas as pd
 import hashlib
-from functools import partial
 from pathlib import Path
 import numpy as np
 import logging
-import inspect
 
 from .meta import subkwargs
 
@@ -322,7 +320,7 @@ def df_from_csvs(csvs, dropna=True, ignore_index=False, **kwargs):
     - Existing neighbours take priority (i.e. FeO won't be inserted bf Al2O3)
     - Earlier inputs take priority (where ordering is ambiguous, place the earlier first)
 
-    Todo
+    Todo:
     -----
     Attempt to preserve column ordering across column sets, assuming
     they are generally in the same order but preserving only some of the
@@ -330,7 +328,7 @@ def df_from_csvs(csvs, dropna=True, ignore_index=False, **kwargs):
     """
     cols = []
     dfs = []
-    for ix, t in enumerate(csvs):
+    for t in csvs:
         dfs.append(pd.read_csv(t, **kwargs))
         cols = cols + [i for i in dfs[-1].columns if i not in cols]
 
