### 2020-02-24 bea5f8b
Whitespace removal, clean up for mineral.template
