+++ docs/source/api/geochem/ions.rst
@@ -1,0 +1,5 @@
+pyrolite\.geochem\.ions
+-------------------------------
+  .. automodule:: pyrolite.geochem.ions
+      :members:
+      :undoc-members:
