### 2020-02-24 288fc85
Update geochem.ions docstring
