### 2020-02-25 3577a0d
Minor cleanup for util.skl.impute
