+++ test/util/skl/util_skl_impute.py
@@ -1,17 +1,18 @@
 import unittest
+import numpy as np
 from pyrolite.util.synthetic import test_df
 from pyrolite.comp.codata import close
-
 
 try:
     import sklearn
     from sklearn.svm import SVC
     from sklearn.model_selection import GridSearchCV
+
     HAVE_SKLEARN = True
 
     def test_classifier():
         param_grid = dict(gamma=np.array([0.001, 0.01]), C=np.array([1, 10]))
-        gs = GridSearchCV(SVC(gamma='scale'), param_grid, cv=2)
+        gs = GridSearchCV(SVC(gamma="scale"), param_grid, cv=2)
         return gs
 
 
