### 2020-02-25 e589150
Bugfix imports for util.plot
