### 2020-02-25 c30ce3c
Clean up after util.plot update
