+++ pyrolite/plot/density/ternary.py
@@ -4,7 +4,6 @@ import matplotlib.tri
 from ...comp.codata import close, inverse_ilr, ilr, alr, inverse_alr
 from ...util.math import flattengrid
 from ...util.distributions import sample_kde
-from ...util.plot.axes import axes_to_ternary
 from ...util.plot.grid import bin_centres_to_edges
 import logging
 
