+++ pyrolite/util/plot/grid.py
@@ -3,6 +3,8 @@ Gridding and binning functions.
 """
 import numpy as np
 from ...comp.codata import close
+from .transform import ABC_to_xy, xy_to_ABC
+import scipy.interpolate
 import logging
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
