### 2020-02-25 3632b60
Remove duplicate import for plot
Whitespace removal for util.plot.legend
