+++ pyrolite/util/skl/impute.py
@@ -8,21 +8,16 @@ logger = logging.getLogger(__name__)
 
 try:
     from sklearn.base import TransformerMixin, BaseEstimator
+    from sklearn.import import IterativeImputer
 except ImportError:
     msg = "scikit-learn not installed"
-    logger.warning(msg)
-
-try:
-    from fancyimpute import IterativeImputer, SoftImpute
-except ImportError:
-    msg = "fancyimpute not installed"
     logger.warning(msg)
 
 
 class MultipleImputer(BaseEstimator, TransformerMixin):
     def __init__(self, multiple=5, n_iter=10, groupby=None, *args, **kwargs):
         """
-        Multiple Imputation via :class:`fancyimpute.IterativeImputer`.
+        Multiple Imputation via :class:`sklearn.IterativeImputer`.
 
         Parameters
         ----------
@@ -116,96 +111,5 @@ class MultipleImputer(BaseEstimator, TransformerMixin):
             logger.info(msg)
             for ix in range(self.multiple):
                 self.imputers[ix].fit(X)
-
-        return self
-
-
-class PdSoftImputer(BaseEstimator, TransformerMixin):
-    def __init__(self, max_iters=100, groupby=None, donotimpute=[], *args, **kwargs):
-        """
-        Multiple Imputation via :class:`fancyimpute.SoftImpute`.
-
-        Parameters
-        ----------
-        max_iters : :class:`int`
-            Maximum number of iterations for each imputation.
-        groupby : :class:`str`
-            Column to group by to impute each group separately.
-        donotimpute : :class:`list`
-            Columns not to impute.
-        """
-        self.args = args
-        self.kwargs = kwargs
-        self.max_iters = max_iters
-        self.groupby = groupby
-        self.donotimpute = donotimpute
-
-    def transform(self, X, *args, **kwargs):
-        """
-        Impute Missing Values
-
-        Todo:
-        -----
-            * Need to use masks to avoid :class:`fancyimpute.SoftImpute` returning 0. where it cannot impute.
-        """
-        assert isinstance(X, pd.DataFrame)
-        df = pd.DataFrame(columns=X.columns, index=X.index)  # df of nans
-        df.loc[:, self.donotimpute] = X.loc[:, self.donotimpute]
-        to_impute = [i for i in X.columns if not i in self.donotimpute]
-        imputable = ~pd.isnull(X.loc[:, to_impute]).all(axis=1)
-        if isinstance(self.imputer, dict):
-            for d in self.imputer.values():
-                mask = d["mask"]
-                mask = mask & imputable
-                imputer = d["impute"]
-                imputed_data = imputer.fit_transform(X.loc[mask, to_impute])
-                assert imputed_data.shape[0] == X.loc[mask, :].index.size
-                df.loc[mask, to_impute] = imputed_data
-
-            return df
-        else:
-            imputed_data = self.imputer.fit_transform(X.loc[imputable, to_impute])
-            assert imputed_data.shape[0] == X.loc[imputable, :].index.size
-            df.loc[imputable, to_impute] = imputed_data
-            return df
-
-    """
-    def inverse_transform(self, Y, *args, **kwargs):
-        # For non-compositional data, take the mask and reverting to nan
-        # for compositional data, renormalisation would be needed
-        pass
-    """
-
-    def fit(self, X, y=None):
-        assert isinstance(X, pd.DataFrame)
-        # start = X
-        y_present = y is not None
-        groupby_present = self.groupby is not None
-        self.imputer = []
-        if y_present or groupby_present:
-            assert not (groupby_present and y_present)
-            if y_present:
-                classes = np.unique(y)
-                gen_mask = lambda c: y == c
-            if groupby_present:
-                classes = X[self.groupby].unique()
[…]
